Health bars look decent, attack range is fixed, game got a small reformat, bases are now circles and a bunch of other smaller stuff.

+++ app/src/main/java/com/ede/standyourground/framework/api/LatLngService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface LatLngService {
 
-    float bearing(LatLng to, LatLng from);
+    double bearing(LatLng to, LatLng from);
 
     LatLng midpoint(LatLng l1, LatLng l2);
